Prevent deletion of certification type if certifications exist

+++ app/services/certification_type_service.rb
@@ -31,6 +31,10 @@ class CertificationTypeService
   end
 
   def delete_certification_type(certification_type)
+    if certification_type.certifications.any?
+      return :certification_exists
+    end
+
     certification_type.destroy
   end
